+++ lib/chef/provider/one_vnet_lease.rb
@@ -1,7 +1,22 @@
+# Copyright 2015, BlackBerry, Inc.
+#
+#Licensed under the Apache License, Version 2.0 (the "License");
+#you may not use this file except in compliance with the License.
+#You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+#Unless required by applicable law or agreed to in writing, software
+#distributed under the License is distributed on an "AS IS" BASIS,
+#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#See the License for the specific language governing permissions and
+#limitations under the License.
 
 class Chef::Provider::OneVnetLease < Chef::Provider::LWRPBase
   use_inline_resources
 
+  provides :one_vnet_lease
+  
   attr :current_vnet
   
   def action_handler
@@ -10,10 +25,7 @@ class Chef::Provider::OneVnetLease < Chef::Provider::LWRPBase
 
   def exists?
     new_driver = get_driver
-    filter = {
-      :id => @new_resource.vnet.is_a?(Integer) ? @new_resource.vnet : nil, 
-      :name => @new_resource.vnet.is_a?(String) ? @new_resource.vnet : nil
-    }
+    filter = { @new_resource.vnet.is_a?(Integer) ? :id : :name => @new_resource.vnet }
     @current_vnet = new_driver.one.get_resource('vnet', filter)
     raise "vnet '#{@new_resource.vnet}' does not exist" if @current_vnet.nil?
     @current_vnet.info!